Use toInt rather than toLong when unicode escaping.

+++ src/main/scala/argonaut/PrettyParams.scala
@@ -124,7 +124,7 @@ object StringEscaping {
       case '\n' => "\\n"
       case '\r' => "\\r"
       case '\t' => "\\t"
-      case possibleUnicode if possibleUnicode.isControl => "\\u%04x".format(possibleUnicode.toLong)
+      case possibleUnicode if possibleUnicode.isControl => "\\u%04x".format(possibleUnicode.toInt)
       case _ => c.toString
     }
 }
